requirements.txt
python-docx
beautifulsoup4
flask
openpyxl
pandas